pyproject.toml
[tool.poetry]
name = "chat-langchain"
version = "0.1.1"
description = ""
authors = ["SN <6432132+samnoyes@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "0.3.16"
langsmith =  "^0.1.25"
langchain-community = "0.3.16"
langchain-openai = "0.3.3"
langchain-google-genai = "2.0.0"
langchain-anthropic = "0.2.0"
langchain-fireworks = "0.2.0"
langserve = "0.3.1"
fastapi = "^0.103.1"
pydantic = "2.7.4"
uvicorn = "^0.23.2"
beautifulsoup4 = "^4.12.2"
weaviate-client = ">=3.26.7,<4.0.0"
psycopg2 = "^2.9.7"
lxml = "^4.9.3"
voyageai = "^0.1.4"
pillow = "^10.2.0"
tiktoken = "0.7"
cohere = "4.45"
auth = "^0.5.3"
langchain-core = "0.3.33"
langchain-deepseek = "0.1.1"
langchain-zhipuai = "0.0.1"
langchainhub = "0.1.21"
langchain-text-splitters = "0.3.3"
sse-starlette = "1.3.3"
zhipuai = "^2.1.5.20250801"
ragas = "^0.3.0"
beir = "^2.2.0"
tqdm = "^4.67.1"
jq = "^1.10.0"


[tool.poetry.group.lint.dependencies]
ruff = "^0.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
